APISage/requirements.txt
langchain>=0.2.0
langgraph>=0.1.0
langchain-community>=0.2.0
ollama>=0.2.0
llama-cpp-python>=0.2.0
transformers>=4.38.0
torch>=2.2.0
fastapi>=0.110.0
uvicorn>=0.27.0
redis>=5.0.0
asyncpg>=0.29.0
pydantic>=2.5.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
playwright>=1.42.0
qdrant-client>=1.8.0
milvus>=2.3.0
prometheus-client>=0.19.0
structlog>=24.1.0
tenacity>=8.2.0
ragas>=0.1.0
deepeval>=0.20.0
duckduckgo-search>=5.0.0
trafilatura>=1.8.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
sentence-transformers>=2.5.0

# Multi-LLM Provider Support
litellm>=1.0.0

# Document Parsers
postmanparser>=0.1.0
unstructured>=0.10.0
pyyaml>=6.0.0
python-multipart>=0.0.6

# Additional Vector Databases
chromadb>=0.4.0
pymilvus>=2.3.0
pinecone-client>=3.0.0

# Enhanced API Processing
jsonschema>=4.20.0
pydantic[email]>=2.5.0

# Search and Ranking
rank-bm25>=1.2.2
